backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy==1.24.4
scikit-learn==1.3.2
pillow==10.2.0
tqdm==4.66.1
python-dotenv==1.0.0
celery==5.3.4
httpx==0.25.2
tenacity==8.2.3
h5py==3.10.0
pandas==2.1.4

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2